# internal/docs/swagger.yaml
basePath: /api
definitions:
...
  answers.createMediaRequest:
    properties:
      answer_id:
        type: string
      checksum:
        type: string
      duration_seconds:
        type: integer
      file_key:
        type: string
      mime_type:
        type: string
      size_bytes:
        type: integer
    required:
    - file_key
    - mime_type
    - size_bytes
    type: object
  answers.mediaResponse:
    properties:
      answer_id:
        type: string
      checksum:
        type: string
      created_at:
        type: string
      duration_seconds:
        type: integer
      file_key:
        type: string
      id:
        type: string
      mime_type:
        type: string
      participant_id:
        type: string
      project_id:
        type: string
      size_bytes:
        type: integer
      status:
        type: string
      updated_at:
        type: string
    type: object
...
  projects.projectResponse:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      owner_id:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  projects.updateProjectRequest:
    properties:
      description:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API with Clean Architecture
  title: Indagio API
  version: "1.0"
paths:
...
  /projects/{id}:
    delete:
      description: Delete a project permanently
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete project
      tags:
      - Projects
    get:
      description: Get one project by ID when the authenticated user has access
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/projects.projectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get project by ID
      tags:
      - Projects
    patch:
      consumes:
      - application/json
      description: Update project name, description, or status
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Project patch payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/projects.updateProjectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/projects.projectResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update project
      tags:
      - Projects
  /projects/{id}/answers:
    post:
      consumes:
      - application/json
      description: Create a participant answer record in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Answer payload
        in: body
        name: body
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create answer
      tags:
      - Answers
...
  /projects/{id}/participants/{participantID}/answers:
    get:
      description: List answers submitted by a participant in a project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant ID
        in: path
        name: participantID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List participant answers
      tags:
      - Answers
  /projects/{id}/participants/{participantID}/media:
    post:
      consumes:
      - application/json
      description: Register uploaded participant media metadata
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: string
      - description: Participant ID
        in: path
        name: participantID
        required: true
        type: string
      - description: Media payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/answers.createMediaRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/answers.mediaResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "422":
          description: Unprocessable Entity
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create media record
      tags:
      - Answers
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"